Change on the equantions in gp2
We created six new variables to make the code more readable and
to be more consistent with the reference (McMorran 2008). They are
argument_d, argument_f, argument_p, argument_v that are the arguments
of the exponential functions 18(b,c,d and e) from the reference.
The two other variables function_d and function_v are the evaluation
of the exponential in 18b and 18e. The complex exponetials are
being evaluated using Euler's identity through the cossine and sine of
phiI[i] in line 284. All the equations now match with the article,
unlike before, that had some typos and some terms from McMorran's code
that we do not understand the origin nor their meaning (for instance,
G2_x where McMorran himself does not show its origin).