New timers added
The first timer was changed to compute the entire time spent in gp2.
Two new timers were added to both gp0 and gp1 as well.
In central simulator it was added a printf to each time the program
enter one of the gp's, printing the 'i' it is inthe moment.
All printf's are as comments.