Removed IF statements that consider electron interactions
In the previous code, the Gratings.c file had an IF statement where it chose to
consider image charge effects between the electron beam and the grating (basically
a Coulomb like interaction that cause a phase shift to the electron wave). Since
we decided not to model the beam of charged particles for now, we removed the if
statement and went straigth to the declaration of the variable for the atom beam case.