Changing abzloc = sp.height to abszloc = current_z_position
Inside the intensity_after_1st_grating function, abzloc was receiving the value
of sp.height. We believe this is wrong: the same variable was used inside
intensity_after_2nd_grating and it received the sp.current_z_position value.
 Furthermore, receiving the value of sp.height does not seem to make sense
because height is a dimension on the x direction, not in the z direction (if we
understood this correctly, this should be the size of the grating, approx 50nm).