New variable "account_Van_der_Waals" + IF statement for sinc() coef reimplemented
We realized that the if statements that were removed in the commit named
"Removed IF statements that consider electron interactions" were still necessary
in case we wanted to run things without using gravity or Van der Waals interaction.
Therefore, we created a variable "account_Van_der_Waals" and implemented it to have
its value chosen in the command line. That being done, we recreated the IF statements
inside functions "instensity_after_1st_grating" and  "instensity_after_2nd_grating"
but now the conditional is not "useimagecharge == 0" anymore, instead it is
"account_gravity == 1 || account_Van_der_Waals == 1".